API tests refactored
